Move `handleAnchor` functions to `helpers`
