+++ src/components/about/BackgroundEducation.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { handleAnchorAbout } from "../../lib/handleAnchorAbout";
+  import { handleAnchorAbout } from "../../helpers/handleAnchorAbout";
 </script>
 
 <p>
